+++ src/c4group/CStdFile.h
@@ -37,7 +37,7 @@ public:
 	virtual bool Read(void *pBuffer, size_t iSize) = 0;
 	virtual bool Advance(int iOffset) = 0;
 	// Get size. compatible with c4group!
-	virtual int AccessedEntrySize() = 0;
+	virtual size_t AccessedEntrySize() = 0;
 	virtual ~CStdStream() {}
 };
 
@@ -77,7 +77,7 @@ public:
 	          bool fCompressed = false);
 	// flush contents to disk
 	inline bool Flush() { if (ModeWrite && BufferLoad) return SaveBuffer(); else return true; }
-	int AccessedEntrySize();
+	size_t AccessedEntrySize();
 protected:
 	void ClearBuffer();
 	int LoadBuffer();
